Backend/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Deprecated==1.2.13
dnspython==2.3.0
ecdsa==0.18.0
Faker==17.6.0
fastapi==0.92.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jose==1.0.0
kafka-python==2.0.2
MarkupSafe==2.1.2
matcher==0.2
matchers==0.22
mongoengine==0.26.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
PyHamcrest==2.0.4
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
websockets==10.4
wrapt==1.14.1